requirements.txt
python-dotenv
langchain
langchain-openai
langchain-core
langchain-community
langchain-huggingface
langchain-groq
Pymupdf
Wikipedia
Arxiv
bs4
faiss-cpu
lxml
langchain-huggingface
pypdf
langgraph
langgraph-cli[inmem]
pandas
pydantic
langchain-mongodb
pymongo
mysql-connector-python